pyproject.toml
[project]
name = "swift"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aerich>=0.9.0",
    "bcrypt==3.2.0",
    "fastapi>=0.115.12",
    "orjson>=3.10.18",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.11.4",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "tomlkit>=0.13.2",
    "tortoise-orm>=0.25.0",
    "uvicorn>=0.34.2",
]

[tool.aerich]
tortoise_orm = "database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
